# server/openapi.yaml
openapi: 3.1.0
info:
  title: QFree API
  version: 1.0.0
  description: |
    Smart healthcare queue management and real-time waiting-time information.

    **Auth:** send `Authorization: Bearer <accessToken>` (15 min). Renew with `POST /auth/refresh`,
    which reads the httpOnly `qf_rt` cookie and rotates it.

    **Errors:** `{ "error": { "code": "QUEUE_FULL", "message": "…", "details": … } }`.

    **Real-time:** Socket.IO on the same origin. Handshake `auth: { token }` (optional).
    Client → server: `queue:watch`, `queue:unwatch`, `queue:watch-staff`, `queue:unwatch-staff`, `org:watch` (arg: id, ack `{ok, error?}`).
    Server → client: `queue:snapshot` (QueueSnapshot), `queue:entry` (EntryView), `queue:staff` (StaffView), `org:queue`, `notification:new`.
servers:
  - url: /api
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
    refreshCookie: { type: apiKey, in: cookie, name: qf_rt }
  parameters:
    id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
    page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    pageSize: { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
...
  responses:
    Error:
      description: Error
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, examples: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, QUEUE_FULL, QUEUE_CLOSED, QUEUE_PAUSED, ALREADY_IN_QUEUE, OUTSIDE_HOURS, DOCTOR_UNAVAILABLE, TOO_MANY_ACTIVE_QUEUES, RATE_LIMITED] }
            message: { type: string }
            details: {}
...
security:
  - bearer: []
paths:
...
  /patients/me:
    get: { tags: [Patients], responses: { '200': { description: OK } } }
    patch: { tags: [Patients], requestBody: { content: { application/json: { schema: { type: object, properties: { fullName: { type: string }, phone: { type: [string, 'null'] }, dateOfBirth: { type: [string, 'null'], format: date }, gender: { type: [string, 'null'], enum: [FEMALE, MALE, OTHER, PREFER_NOT_TO_SAY, null] } } } } } }, responses: { '200': { description: OK } } }
...
  /patients/me/history:
    get: { tags: [Patients], parameters: [{ $ref: '#/components/parameters/page' }, { $ref: '#/components/parameters/pageSize' }], responses: { '200': { description: Paged visits } } }
  /patients/{id}/history:
    get: { tags: [Patients], summary: Self or admin (admin access audited), parameters: [{ $ref: '#/components/parameters/id' }], responses: { '200': { description: Paged visits }, '403': { $ref: '#/components/responses/Error' } } }

  # ───────── Doctors (self-service) ─────────
...